# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DayFlow API
  version: 1.0.0
  description: API REST versionada para autenticación, usuarios, catálogos, tickets,
    auditoría y analítica de DayFlow.
paths:
...
  /api/v1/tickets/{id}/assign/:
    post:
      operationId: tickets_assign_create
      description: Ticket visibility and audited workflow actions.
      summary: Asignar técnico
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este ticket.
        required: true
      tags:
      - Tickets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketAssignmentRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TicketAssignmentRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TicketAssignmentRequest'
        required: true
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketList'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                Validación:
                  value:
                    message: No se pudo procesar la solicitud.
                    fields:
                      due_date:
                      - La fecha límite no puede ser anterior a la creación.
          description: Solicitud o campos inválidos.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                NoAutenticado:
                  value:
                    message: Las credenciales de autenticación no se proveyeron.
                    fields: {}
                  summary: No autenticado
          description: JWT ausente, inválido o expirado.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                SinPermiso:
                  value:
                    message: No tienes permisos para realizar esta operación.
                    fields: {}
                  summary: Sin permiso
          description: El rol o el estado de la cuenta no permite la operación.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                NoEncontrado:
                  value:
                    message: No encontrado.
                    fields: {}
                  summary: No encontrado
          description: El recurso no existe o queda fuera del scope del usuario.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                Conflicto:
                  value:
                    message: El ticket ya fue tomado por otro técnico.
                    fields: {}
          description: Conflicto con el estado actual del ticket.
  /api/v1/tickets/{id}/attachments/:
    post:
      operationId: tickets_attachments_create
      description: Ticket visibility and audited workflow actions.
      summary: Cargar adjunto
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este ticket.
        required: true
      tags:
      - Tickets
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TicketAttachmentUploadRequest'
        required: true
      security:
      - bearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketAttachment'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                Validación:
                  value:
                    message: No se pudo procesar la solicitud.
                    fields:
                      due_date:
                      - La fecha límite no puede ser anterior a la creación.
          description: Solicitud o campos inválidos.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                NoAutenticado:
                  value:
                    message: Las credenciales de autenticación no se proveyeron.
                    fields: {}
                  summary: No autenticado
          description: JWT ausente, inválido o expirado.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                SinPermiso:
                  value:
                    message: No tienes permisos para realizar esta operación.
                    fields: {}
                  summary: Sin permiso
          description: El rol o el estado de la cuenta no permite la operación.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                NoEncontrado:
                  value:
                    message: No encontrado.
                    fields: {}
                  summary: No encontrado
          description: El recurso no existe o queda fuera del scope del usuario.
...
  /api/v1/tickets/{id}/comments/:
    post:
      operationId: tickets_comments_create
      description: Ticket visibility and audited workflow actions.
      summary: Agregar comentario
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este ticket.
        required: true
      tags:
      - Tickets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCommentCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TicketCommentCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TicketCommentCreateRequest'
        required: true
      security:
      - bearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketComment'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                Validación:
                  value:
                    message: No se pudo procesar la solicitud.
                    fields:
                      due_date:
                      - La fecha límite no puede ser anterior a la creación.
          description: Solicitud o campos inválidos.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                NoAutenticado:
                  value:
                    message: Las credenciales de autenticación no se proveyeron.
                    fields: {}
                  summary: No autenticado
          description: JWT ausente, inválido o expirado.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                SinPermiso:
                  value:
                    message: No tienes permisos para realizar esta operación.
                    fields: {}
                  summary: Sin permiso
          description: El rol o el estado de la cuenta no permite la operación.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                NoEncontrado:
                  value:
                    message: No encontrado.
                    fields: {}
                  summary: No encontrado
          description: El recurso no existe o queda fuera del scope del usuario.
  /api/v1/tickets/{id}/history/:
    get:
      operationId: tickets_history_list
      description: Ticket visibility and audited workflow actions.
      summary: Consultar historial auditable
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este ticket.
        required: true
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Número de resultados a devolver por página.
        schema:
          type: integer
      tags:
      - Tickets
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTicketHistoryList'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                NoAutenticado:
                  value:
                    message: Las credenciales de autenticación no se proveyeron.
                    fields: {}
                  summary: No autenticado
          description: JWT ausente, inválido o expirado.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                SinPermiso:
                  value:
                    message: No tienes permisos para realizar esta operación.
                    fields: {}
                  summary: Sin permiso
          description: El rol o el estado de la cuenta no permite la operación.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                NoEncontrado:
                  value:
                    message: No encontrado.
                    fields: {}
                  summary: No encontrado
          description: El recurso no existe o queda fuera del scope del usuario.
  /api/v1/tickets/{id}/status/:
    post:
      operationId: tickets_status_create
      description: Ticket visibility and audited workflow actions.
      summary: Cambiar estado
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este ticket.
        required: true
      tags:
      - Tickets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketStatusRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TicketStatusRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TicketStatusRequest'
        required: true
      security:
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketList'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                Validación:
                  value:
                    message: No se pudo procesar la solicitud.
                    fields:
                      due_date:
                      - La fecha límite no puede ser anterior a la creación.
          description: Solicitud o campos inválidos.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                NoAutenticado:
                  value:
                    message: Las credenciales de autenticación no se proveyeron.
                    fields: {}
                  summary: No autenticado
          description: JWT ausente, inválido o expirado.
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                SinPermiso:
                  value:
                    message: No tienes permisos para realizar esta operación.
                    fields: {}
                  summary: Sin permiso
          description: El rol o el estado de la cuenta no permite la operación.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                NoEncontrado:
                  value:
                    message: No encontrado.
                    fields: {}
                  summary: No encontrado
          description: El recurso no existe o queda fuera del scope del usuario.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayFlowError'
              examples:
                Conflicto:
                  value:
                    message: El ticket ya fue tomado por otro técnico.
                    fields: {}
          description: Conflicto con el estado actual del ticket.
...
components:
  schemas:
...
    TicketAssignmentRequest:
      type: object
      properties:
        assigned_technician:
          type: integer
      required:
      - assigned_technician
...
    TicketAttachmentUploadRequest:
      type: object
      properties:
        file:
          type: string
          format: binary
          writeOnly: true
        description:
          type: string
          writeOnly: true
          nullable: true
      required:
      - file
...
    TicketList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          readOnly: true
        description:
          type: string
          readOnly: true
        category:
          type: integer
          readOnly: true
        category_name:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          readOnly: true
        status_label:
          type: string
          readOnly: true
        priority:
          allOf:
          - $ref: '#/components/schemas/PriorityEnum'
          readOnly: true
        priority_label:
          type: string
          readOnly: true
        requester:
          type: integer
          readOnly: true
        requester_name:
          type: string
          readOnly: true
        assigned_technician:
          type: integer
          readOnly: true
        assigned_technician_name:
          type: string
          readOnly: true
        requester_department:
          type: integer
          readOnly: true
        requester_department_name:
          type: string
          readOnly: true
        due_date:
          type: string
          format: date
          readOnly: true
          nullable: true
        taken_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        closed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - assigned_technician
      - assigned_technician_name
      - category
      - category_name
      - closed_at
      - created_at
      - description
      - due_date
      - id
      - priority
      - priority_label
      - requester
      - requester_department
      - requester_department_name
      - requester_name
      - status
      - status_label
      - taken_at
      - title
      - updated_at
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token obtenido en /api/v1/auth/login/. No usar aquí el refresh
        token HttpOnly.
...